Update extention handler and more extentions
