+++ src/MakeSimple.SharedKernel.Infrastructure/Api/ControllerApiBase.cs
@@ -1,0 +1,30 @@
+﻿using MakeSimple.SharedKernel.Contract;
+using Microsoft.AspNetCore.Mvc;
+using System.Net;
+
+namespace MakeSimple.SharedKernel.Infrastructure.Api
+{
+    public class ControllerApiBase : ControllerBase
+    {
+        public virtual IActionResult ResultDTO(IDataResult result)
+        {
+            if (result == null)
+            {
+                return new NotFoundObjectResult(result);
+            }
+
+            return result.StatusCode switch
+            {
+                HttpStatusCode.OK => Ok(result),
+                HttpStatusCode.Conflict => Conflict(result),
+                HttpStatusCode.NoContent => NoContent(),
+                HttpStatusCode.BadRequest => BadRequest(result),
+                HttpStatusCode.NotFound => NotFound(result),
+                _ => new ObjectResult(result)
+                {
+                    StatusCode = (int)result.StatusCode
+                }
+            };
+        }
+    }
+}
