+++ src/MakeSimple.SharedKernel.Infrastructure/Extensions/AppBuilderExtensions.cs
@@ -1,5 +1,13 @@
-﻿using MakeSimple.SharedKernel.Infrastructure.AspNet;
+﻿using MakeSimple.SharedKernel.Contract;
+using MakeSimple.SharedKernel.Infrastructure.DTO;
+using MakeSimple.SharedKernel.Infrastructure.Exceptions;
 using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Diagnostics;
+using Microsoft.AspNetCore.Http;
+using Serilog;
+using System;
+using System.Net;
+using System.Text.Json;
 
 namespace MakeSimple.SharedKernel.Infrastructure.Extensions
 {
@@ -7,7 +15,71 @@ namespace MakeSimple.SharedKernel.Infrastructure.Extensions
     {
         public static IApplicationBuilder UseExceptionHandlerCore(this IApplicationBuilder app)
         {
-            return app.UseMiddleware<HandleExceptionMiddleware>();
+            app.Use(async (context, next) =>
+            {
+                context.Request.EnableBuffering();
+                await next.Invoke();
+            });
+
+            app.UseExceptionHandler(errorApp =>
+            {
+                errorApp.Run(async context =>
+                {
+                    IDataResult result;
+
+                    var exceptionHandlerPathFeature = context.Features.Get<IExceptionHandlerPathFeature>();
+                    var exception = exceptionHandlerPathFeature?.Error;
+                    if (exception != null)
+                    {
+                        if (exception is BaseException baseException)
+                        {
+                            result = baseException.DataResult;
+                        }
+                        else
+                        {
+                            result = new Response<bool>
+                            (
+                                HttpStatusCode.InternalServerError,
+                                new ErrorBase()
+                                {
+                                    Code = "InternalServerError",
+                                    ErrorMessage = "Internal Server Error"
+                                }
+                            );
+                        }
+                    }
+                    else
+                    {
+                        exception = new Exception("Unhandled");
+
+                        result = new Response<bool>
+                        (
+                            HttpStatusCode.InternalServerError,
+                            new ErrorBase()
+                            {
+                                Code = "Unhandled",
+                                ErrorMessage = "Unhandled"
+                            }
+                        );
+                    }
+
+                    string jsonResult = JsonSerializer.Serialize(result);
+                    string requestInfo = $"HTTP Request:\n" +
+                    $"\tClaims: {context.User?.CovertToString()}\n" +
+                    $"\tHeaders: {context.Request.Headers?.CovertToString()}\n" +
+                    $"\tBody: {await context.Request.Body?.ReadToEndBufferingAsync()}\n" +
+                    $"\tQueryString: {context.Request.QueryString}\n" +
+                    $"\tResponse: {jsonResult}\n";
+
+                    Log.Error(exception.GetBaseException(), requestInfo);
+
+                    context.Response.StatusCode = (int)result.StatusCode;
+                    context.Response.ContentType = "application/json";
+                    await context.Response.WriteAsync(jsonResult);
+                });
+            });
+
+            return app;
         }
     }
