--- src/MakeSimple.SharedKernel.Infrastructure/Helpers/UuuidHelper.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace MakeSimple.SharedKernel.Infrastructure.Helpers
-{
-    public static class UuuidHelper
-    {
-        public static string GenerateId(string guid = null)
-        {
-            return string.IsNullOrEmpty(guid) ? guid : Guid.NewGuid().ToString();
-        }
-    }
-}
