+++ src/MakeSimple.SharedKernel/Contract/IDataResult.cs
@@ -3,7 +3,7 @@
 namespace MakeSimple.SharedKernel.Contract
 {
     public interface IDataResult
-    {
+    { 
         HttpStatusCode StatusCode { get; set; }
         IError Error { get; set; }
         void CopyFrom(IDataResult source);
